Assignment Of Problem Solving
Problem Statement: Alice and Bob are playing a game on a sequence a1,a2,…,an of length n. They move in turns and Alice moves first.
In the turn of each player, he or she should select an integer and remove it from the sequence. The game ends when there is no integer left in the sequence.
Alice wins if the sum of her selected integers is even; otherwise, Bob wins.
Your task is to determine who will win the game, if both players play optimally.